Make the toolbar respond to a content overflow
The breakpoint defined in ext.gettingstarted.taskToolbar.lowWidth.less
is fixed and therefore doesn't handle long translations well. Instead,
calculate the point of overflow when the toolbar is inserted into the
DOM and monitor the width of the toolbar so that the appropriate
styles can be applied when the breakpoint is reached.

Bug: 61230
Change-Id: I87b89f49f45e65f2f178565ca60a306ed80fbf9f