Add pagination support
Number of discs/pages is set up up front based on the total number of
suggestions.

Pages are rendered on demand, and buttons are enabled/disabled,
corresponding with the discs.

Works for any number of suggestions (0 is already handled, 1 to 9
works with the dynamic pagination)

For IE, show only IE 8+ for now.

Change-Id: I8423168691e8cd9e7a0c6b39d8ffca80195403df