Remove 'active' from personal tool
I think this is meant for cases when the personal tool corresponds
to the page.  For example, when you're on User talk:MyUsername
the 'mytalk' personal tool is active.

Since we don't have a page corresponding to the flyout,
it doesn't currently apply.

(P.S. You can see a potential appearance for this by going to
your talk page in Monobook).

Change-Id: I7b7a7c4aad15d62ab0e239e774bb74c5b3efdbe1