Exclude users with null user_registration
Previously, they were incorrectly considered to have registered
now (the time wfTimestamp was called).

Add corresponding test.  Since this must check that NULL is not
equivalent to now, make it so now would be bucketed into the test.
This requires changing the other tests to be relative to now as well.

Bug: 70759
Change-Id: I73e0315e270f2754199d599074a06cf3fe4daddc