Allow suggestions to be opened in new tabs/windows
If the user ctrl-clicks (meta-clicks on OS X) or middle-clicks a task
recommendation, then the associated TaskRecommendationClick event is
logged without navigation.

At the time of writing, this function handles left/middle/right clicks
in Chrome, Safari and Firefox. Middle and ctrl-clicks aren't handled in
IE8-11, nor are middle clicks in Opera.

Change-Id: I807200e1071c78c1518741961d20a2269c0a9177