Fill required field when bypassing CAPTCHA
I08244add stopped blacklisting the "required" HTML5 form validation
attribute. Consequently, the CAPTCHA on the registration page is now
marked as required and a UA that supports these attributes - like those
we use in our automated browser testing environment - will not allow the
registration form to be submitted without the CAPTCHA being filled.

This happens in spite of using the CAPTCHA bypass mechanism introduced
in Iad735755.

Bug: T155797
Change-Id: I3c1362dba3147ac974cba4f3c379771660b85b38