[instrumentation] Remove old schemas
The SignupExp* schemas were used as part of the "Asking anonymous
editors to register" experiment [0]. The experiment was run circa July
2014.

All registered schema/revision ID pairs are bundled with the
EventLogging client and therefore negatively impact the performance
budget of that module. It's imperative that unused schemas are tidied up
quickly.

[0] https://meta.wikimedia.org/wiki/Research:Asking_anonymous_editors_to_register

Change-Id: I896ef5ef42f61248fa20d3bbe9a8247b76e73afd